Move tests to the same folder where the code is
* Refactor some tests
