+++ core/core_suite_test.go
@@ -1,0 +1,13 @@
+package core_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestSnickers(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Core Suite")
+}
