+++ db/config.json
@@ -1,0 +1,6 @@
+{
+  "SWAP_DIRECTORY": "/tmp/",
+  "LOGFILE": "/tmp/output.log",
+  "PORT": "8000",
+  "MONGODB_HOST": "localhost"
+}
