+++ db/db_suite_test.go
@@ -1,0 +1,13 @@
+package db_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestDb(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Db Suite")
+}
